+++ logcat/src/androidMain/kotlin/com/airsaid/logcat/AndroidLogcatLogger.kt
@@ -1,7 +1,10 @@
 package com.airsaid.logcat
 
+import android.app.Application
+import android.content.pm.ApplicationInfo
 import com.airsaid.logcat.AndroidLogcatLogger.Companion.install
 import com.airsaid.logcat.LogPriority.DEBUG
+import com.airsaid.logcat.internal.isDebuggableApp
 
 /**
  * A [logcat] logger that delegates to [AndroidLogcatLogStrategy] for any log with a priority of
@@ -10,7 +13,9 @@ import com.airsaid.logcat.LogPriority.DEBUG
  * Handles special cases for [LogPriority.ASSERT] (which requires sending to Log.wtf) and
  * splitting logs to be at most 4000 characters per line (otherwise logcat just truncates).
  *
- * Call [install] to make sure you never log in release builds.
+ * Call [installOnDebuggableApp] to install this logger only for debuggable applications. [install]
+ * always installs the logger and should be used only when logging in non-debuggable builds is
+ * intentional.
  *
  * @author airsaid
  */
@@ -28,6 +33,31 @@ class AndroidLogcatLogger(
   }
 
   companion object {
+    /**
+     * Installs an [AndroidLogcatLogger] only when [application] is debuggable.
+     *
+     * @param application the application used to check [ApplicationInfo.FLAG_DEBUGGABLE].
+     * @param minPriority the minimum priority to log.
+     * @param formatStrategy the strategy used to format and write log messages.
+     */
+    fun installOnDebuggableApp(
+      application: Application,
+      minPriority: LogPriority = DEBUG,
+      formatStrategy: FormatStrategy<AndroidLogcatLogStrategy>,
+    ) {
+      if (!application.isDebuggableApp) return
+      install(minPriority, formatStrategy)
+    }
+
+    /**
+     * Installs an [AndroidLogcatLogger] in all build types.
+     *
+     * Prefer [installOnDebuggableApp] when logcat output should be disabled for non-debuggable
+     * applications.
+     *
+     * @param minPriority the minimum priority to log.
+     * @param formatStrategy the strategy used to format and write log messages.
+     */
     fun install(
       minPriority: LogPriority = DEBUG,
       formatStrategy: FormatStrategy<AndroidLogcatLogStrategy>,